+++ src/django/api/models/__init__.py
@@ -79,3 +79,4 @@ from .user import (
 )
 from .version import Version
 from .moderation_event import ModerationEvent
+from .download_location_payment import DownloadLocationPayment